Update Prisma schema for user models
The Prisma schema has been updated with important modifications to the Admin, Advisor, and Investor models. These changes include the addition of an "access_type" attribute to the Admin model and introduction of a "phone_number" attribute to the Advisor and Investor models for extended user information. Moreover, the 'type_name' attribute in the InvestmentType model is now unique to prevent redundancy.